subscription and inventory

+++ AwawaTech.Mecanaut.API/Shared/Infrastructure/Persistence/EFC/Repositories/BaseRepository.cs
@@ -33,7 +33,7 @@ public class BaseRepository<TEntity> : IBaseRepository<TEntity> where TEntity : 
     }
 
     // inheritedDoc
-    public async Task<TEntity?> FindByIdAsync(int id)
+    public async Task<TEntity?> FindByIdAsync(long id)
     {
         return await Context.Set<TEntity>().FindAsync(id);
     }
